+++ Sources/Config/Config+Arguments.swift
@@ -1,5 +1,4 @@
 import Foundation
-import Node
 
 extension Node {
     /**
