+++ Sources/Config/KeyAccessible.swift
@@ -1,0 +1,7 @@
+internal protocol KeyAccessible {
+    associatedtype Key
+    associatedtype Value
+    subscript(key: Key) -> Value? { get set }
+}
+
+extension Dictionary: KeyAccessible {}
