+++ Sources/Config/Source.swift
@@ -1,0 +1,5 @@
+public enum Source {
+    case memory(name: String, config: Node)
+    case commandline
+    case directory(root: String)
+}
